[test] Fix implicit function declarations and remove -fcommon from lit.cfg
Remove -fcommon from the global common_opts in test/lit.cfg so it is no
longer injected into every %clangopts/%clangg0opts/%embedclangopts
substitution.

Two classes of fixes are applied across all targets (Hexagon, RISCV,
RISCV64, ARM, AArch64, X86):

1. Add -fcommon explicitly to the RUN

2. Add forward declarations (-Wimplicit-function-declaration)

3. Casting pointer-to-integer conversions that became errors under -Wint-conversion

+++ test/Common/standalone/ProvideStandardSymbols/Inputs/1.c
@@ -1,4 +1,4 @@
 extern int __ehdr_start;
 int foo() {
-  return &__ehdr_start;
+  return (int)(long)(&__ehdr_start);
 }
